% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{build_edges}
\alias{build_edges}
\title{Build edges from parse data}
\usage{
build_edges(parse_data)
}
\arguments{
\item{parse_data}{Named list from parse() function}
}
\value{
Data frame with 'from' and 'to' columns representing edges
}
\description{
Build edges from parse data
}
\keyword{internal}
