% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{build_edges_and_adjacency}
\alias{build_edges_and_adjacency}
\title{Build edges and adjacency list from parse data}
\usage{
build_edges_and_adjacency(parse_data, nodes)
}
\arguments{
\item{parse_data}{Named list from parse() function}

\item{nodes}{Character vector of all nodes}
}
\value{
List with 'edges' data frame and 'adjacency_list'
}
\description{
Build edges and adjacency list from parse data
}
\keyword{internal}
