% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{collect_all_nodes}
\alias{collect_all_nodes}
\title{Collect all unique nodes from parse data}
\usage{
collect_all_nodes(parse_data)
}
\arguments{
\item{parse_data}{Named list from parse() function}
}
\value{
Character vector of all unique nodes
}
\description{
Collect all unique nodes from parse data
}
\keyword{internal}
