% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{compute_file_checksum}
\alias{compute_file_checksum}
\title{Compute MD5 checksum for a file}
\usage{
compute_file_checksum(file_path)
}
\arguments{
\item{file_path}{Path to the file}
}
\value{
Character string containing the MD5 checksum
}
\description{
Compute MD5 checksum for a file
}
\keyword{internal}
