% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{compute_stale_nodes}
\alias{compute_stale_nodes}
\title{Compute stale nodes using reworked algorithm}
\usage{
compute_stale_nodes(graph_obj, state_obj, parse_data)
}
\arguments{
\item{graph_obj}{Graph object with nodes and edges}

\item{state_obj}{State object from read_state()}

\item{parse_data}{Parse data for understanding script dependencies}
}
\value{
Character vector of stale script names
}
\description{
Implements the reworked staleness marking algorithm:
\enumerate{
\item Get list of all fresh files from state_obj
\item Set stale nodes as nodes that are affected by non-fresh files
\item Iterate stale nodes, mark descendants as stale
}
}
\keyword{internal}
