% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{detect_cycles}
\alias{detect_cycles}
\title{Detect cycles in the dependency graph using DFS}
\usage{
detect_cycles(graph_obj)
}
\arguments{
\item{graph_obj}{Graph object from graph() function}
}
\description{
Detect cycles in the dependency graph using DFS
}
\keyword{internal}
