% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{detect_cycles_with_df}
\alias{detect_cycles_with_df}
\title{Detect cycles in the dependency graph using DFS (with data frame nodes)}
\usage{
detect_cycles_with_df(graph_obj)
}
\arguments{
\item{graph_obj}{Graph object from graph() function}
}
\description{
Detect cycles in the dependency graph using DFS (with data frame nodes)
}
\keyword{internal}
