% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{determine_node_staleness}
\alias{determine_node_staleness}
\title{Determine staleness for script nodes based on their dependencies}
\usage{
determine_node_staleness(script_names, parse_data, state_obj = NULL)
}
\arguments{
\item{script_names}{Character vector of script names}

\item{parse_data}{Named list from parse() function}

\item{state_obj}{Optional state object from read_state()}
}
\value{
Logical vector indicating staleness for each script
}
\description{
Determine staleness for script nodes based on their dependencies
}
\keyword{internal}
