% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{determine_staleness}
\alias{determine_staleness}
\title{Determine staleness for files}
\usage{
determine_staleness(files, state_obj = NULL)
}
\arguments{
\item{files}{Character vector of file names}

\item{state_obj}{Optional state object from read_state()}
}
\value{
Logical vector indicating staleness
}
\description{
Determine staleness for files
}
\keyword{internal}
