% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{display_scripts_table_targets}
\alias{display_scripts_table_targets}
\title{Display the scripts table using targets backend}
\usage{
display_scripts_table_targets(pipeline_data)
}
\arguments{
\item{pipeline_data}{Parsed pipeline data}
}
\description{
Display the scripts table using targets backend
}
