% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{find_parent_script}
\alias{find_parent_script}
\title{Find the parent script that produces a given artifact}
\usage{
find_parent_script(artifact_name, graph_obj)
}
\arguments{
\item{artifact_name}{Name of the artifact file}

\item{graph_obj}{Graph object with edges}
}
\value{
Script name that produces the artifact, or NULL if not found
}
\description{
Find the parent script that produces a given artifact
}
\keyword{internal}
