% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_fresh_files}
\alias{get_fresh_files}
\title{Extract fresh files from state object}
\usage{
get_fresh_files(state_obj)
}
\arguments{
\item{state_obj}{State object from read_state()}
}
\value{
Character vector of fresh file names
}
\description{
Helper function to get list of fresh files from the new state format.
}
\keyword{internal}
