% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_stale_files}
\alias{get_stale_files}
\title{Extract stale files from state object}
\usage{
get_stale_files(state_obj)
}
\arguments{
\item{state_obj}{State object from read_state()}
}
\value{
Character vector of stale file names
}
\description{
Helper function to get list of stale files from the new state format.
}
\keyword{internal}
