% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_script}
\alias{parse_script}
\title{Parse a single R script for file dependencies}
\usage{
parse_script(script_path)
}
\arguments{
\item{script_path}{Absolute path to the R script}
}
\value{
List with 'inputs' and 'outputs' character vectors
}
\description{
Parse a single R script for file dependencies
}
\keyword{internal}
