% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{scripts}
\alias{scripts}
\title{Find all R scripts in project}
\usage{
scripts()
}
\value{
Character vector of absolute paths to .R files
}
\description{
Returns a vector with paths to all .R scripts in or below the project root.
The project root is determined by bakepipe::root().
}
\examples{
\dontrun{
# Get all R scripts in the project
all_scripts <- scripts()
}
}
