% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{topological_sort}
\alias{topological_sort}
\title{Topological sort of the dependency graph}
\usage{
topological_sort(graph_obj)
}
\arguments{
\item{graph_obj}{Graph object from graph() function}
}
\value{
Character vector of nodes in topological order
}
\description{
Returns nodes in topological order using Kahn's algorithm. Scripts will
appear in an order where all dependencies come before the script.
}
\examples{
\dontrun{
parsed <- parse()
graph_obj <- graph(parsed)
execution_order <- topological_sort(graph_obj)
}
}
