% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{validate_external_files}
\alias{validate_external_files}
\title{Validate that external files exist}
\usage{
validate_external_files(graph_obj, parse_data)
}
\arguments{
\item{graph_obj}{Graph object from graph() function}

\item{parse_data}{Parse result with scripts, inputs, outputs, externals}
}
\description{
Ensures that all external files referenced by external_in() calls
actually exist on the filesystem.
}
\keyword{internal}
