% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{validate_input_dependencies}
\alias{validate_input_dependencies}
\title{Validate that input nodes have corresponding producers}
\usage{
validate_input_dependencies(graph_obj, parse_data)
}
\arguments{
\item{graph_obj}{Graph object from graph() function}

\item{parse_data}{Parse result with scripts, inputs, outputs}
}
\description{
Ensures that every input node (files referenced by file_in()) has incoming edges,
meaning they are produced by some script in the pipeline. External files
marked with external_in() are not included in input nodes.
}
\keyword{internal}
