% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{validate_single_producer}
\alias{validate_single_producer}
\title{Validate that each artifact has at most one producer}
\usage{
validate_single_producer(parse_data)
}
\arguments{
\item{parse_data}{Named list from parse() function}
}
\description{
Validate that each artifact has at most one producer
}
\keyword{internal}
