% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Show pipeline status}
\usage{
status()
}
\description{
Display a textual representation of the input/output relationships
between files and artifacts in the console, including Fresh/Stale status
for incremental builds.
}
